feat: add write_env helper, add generate password for env command
